Move utils package into main package
