+++ .changeset/shy-tools-poke.md
@@ -1,0 +1,6 @@
+---
+'@keystone-next/keystone': minor
+'@keystone-next/utils': major
+---
+
+Moved `@keystone-next/utils` to `@keystone-next/keystone/fields/types/image/utils` for image ref related utilities and `@keystone-next/keystone/fields/types/file/utils` for file ref related utilities.
