+++ packages/keystone/package.json
@@ -33,7 +33,6 @@
     "@hapi/iron": "^6.0.0",
     "@keystone-next/admin-ui-utils": "^5.0.6",
     "@keystone-next/types": "^24.0.0",
-    "@keystone-next/utils": "^1.0.4",
     "@keystone-ui/button": "^5.0.0",
     "@keystone-ui/core": "^3.1.1",
     "@keystone-ui/fields": "^4.1.2",
@@ -153,7 +152,8 @@
       "admin-ui/image.tsx",
       "fields/index.ts",
       "fields/types/*/views/index.tsx",
-      "fields/types/relationship/views/RelationshipSelect.tsx"
+      "fields/types/relationship/views/RelationshipSelect.tsx",
+      "fields/types/{image,file}/utils.ts"
     ]
   },
   "repository": "https://github.com/keystonejs/keystone/tree/master/packages/keystone"
