+++ packages/keystone/src/fields/types/file/utils.ts
@@ -1,0 +1,16 @@
+import { AssetMode } from '@keystone-next/types';
+
+const FILEREGEX = /^(local|keystone-cloud):file:([^\\\/:\n]+)/;
+
+export const getFileRef = (mode: AssetMode, name: string) => `${mode}:file:${name}`;
+export const parseFileRef = (ref: string) => {
+  const match = ref.match(FILEREGEX);
+  if (match) {
+    const [, mode, filename] = match;
+    return {
+      mode: mode as AssetMode,
+      filename: filename as string,
+    };
+  }
+  return undefined;
+};
