+++ packages/keystone/src/fields/types/image/index.ts
@@ -9,9 +9,9 @@ import {
   KeystoneContext,
   schema,
 } from '@keystone-next/types';
-import { getImageRef, SUPPORTED_IMAGE_EXTENSIONS } from '@keystone-next/utils';
 import { FileUpload } from 'graphql-upload';
 import { resolveView } from '../../resolve-view';
+import { getImageRef, SUPPORTED_IMAGE_EXTENSIONS } from './utils';
 
 export type ImageFieldConfig<TGeneratedListTypes extends BaseGeneratedListTypes> =
   CommonFieldConfig<TGeneratedListTypes> & {
