+++ packages/keystone/src/fields/types/image/utils.ts
@@ -1,0 +1,37 @@
+import { AssetMode, ImageExtension } from '@keystone-next/types';
+
+const IMAGEREGEX = /^(local|keystone-cloud):image:([^\\\/:\n]+)\.(gif|jpg|png|webp)$/;
+const FILEREGEX = /^(local|keystone-cloud):file:([^\\\/:\n]+)/;
+
+export const getFileRef = (mode: AssetMode, name: string) => `${mode}:file:${name}`;
+export const parseFileRef = (ref: string) => {
+  const match = ref.match(FILEREGEX);
+  if (match) {
+    const [, mode, filename] = match;
+    return {
+      mode: mode as AssetMode,
+      filename: filename as string,
+    };
+  }
+  return undefined;
+};
+
+export const getImageRef = (mode: AssetMode, id: string, extension: ImageExtension) =>
+  `${mode}:image:${id}.${extension}`;
+
+export const SUPPORTED_IMAGE_EXTENSIONS = ['jpg', 'png', 'webp', 'gif'];
+
+export const parseImageRef = (
+  ref: string
+): { mode: AssetMode; id: string; extension: ImageExtension } | undefined => {
+  const match = ref.match(IMAGEREGEX);
+  if (match) {
+    const [, mode, id, ext] = match;
+    return {
+      mode: mode as AssetMode,
+      id,
+      extension: ext as ImageExtension,
+    };
+  }
+  return undefined;
+};
