+++ tests/admin-ui-tests/package.json
@@ -25,7 +25,6 @@
   },
   "dependencies": {
     "@keystone-next/types": "^24.0.0",
-    "@keystone-next/utils": "^1.0.4",
     "@manypkg/find-root": "^1.1.0",
     "dotenv": "^10.0.0",
     "tree-kill": "^1.2.2"
