+++ tests/api-tests/package.json
@@ -27,7 +27,6 @@
   },
   "dependencies": {
     "@keystone-next/types": "^24.0.0",
-    "@keystone-next/utils": "^1.0.4",
     "apollo-cache-control": "^0.14.0",
     "express": "^4.17.1"
   }
